Finishing up FASTAPI

+++ api/routers/plant_logs.py
@@ -20,7 +20,10 @@ def create_plant_log(
     if plant is None:
         raise HTTPException(status_code=404, detail="Plant not found")
     if plant["account_id"] != account_data["id"]:
-        raise HTTPException(status_code=403, detail="Logged in user does not match plant account id")
+        raise HTTPException(
+            status_code=403,
+            detail="Logged in user does not match plant account id",
+        )
     return queries.create(plant_id=plant_id, plant_log_in=plant_log_in)
 
 
@@ -47,28 +50,46 @@ def get_plant_log(
     return plant_log
 
 
-@router.put("/api/plants/plants-logs/{plant_log_id}", response_model=PlantLog)
+@router.put(
+    "/api/plants/{plant_id}/plants-logs/{plant_log_id}",
+    response_model=PlantLog,
+)
 def update_plant_log(
     plant_log_id: str,
+    plant_id: str,
     plant_log_in: PlantLogIn,
     account_data: dict = Depends(authenticator.get_current_account_data),
     queries: PlantLogQueries = Depends(),
+    plant_queries: PlantQueries = Depends(),
 ):
-    return queries.update_plant_log(plant_log_id=plant_log_id, plant_log_in=plant_log_in)
+    plant = plant_queries.get_plant(plant_id)
+    if plant["account_id"] != account_data["id"]:
+        raise HTTPException(
+            status_code=403,
+            detail="Logged in user does not match plant account id",
+        )
+
+    return queries.update_plant_log(
+        plant_log_id=plant_log_id, plant_log_in=plant_log_in
+    )
 
 
-@router.delete("/api/plants/plants-logs/{plant_log_id}", response_model=DeleteStatus)
+@router.delete(
+    "/api/plants/{plant_id}/plants-logs/{plant_log_id}",
+    response_model=DeleteStatus,
+)
 def delete_plant_log(
     plant_log_id: str,
+    plant_id: str,
     account_data: dict = Depends(authenticator.get_current_account_data),
     queries: PlantLogQueries = Depends(),
+    plant_queries: PlantQueries = Depends(),
 ):
+    plant = plant_queries.get_plant(plant_id)
+    if plant["account_id"] != account_data["id"]:
+        raise HTTPException(
+            status_code=403,
+            detail="Logged in user does not match plant account id",
+        )
+
     return {"success": queries.delete_plant_log(plant_log_id)}
-
-
-# @router.get("/api/plants/{plant_id}", response_model=Plant)
-# def get_plant(plant_id: str, queries: PlantQueries = Depends()):
-#     plant = queries.get_plant(plant_id)
-#     if plant is None:
-#         raise HTTPException(status_code=404, detail="Plant not found")
-#     return plant
